requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.7
coverage==7.13.5
Django==6.0.4
django-htmx==1.27.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
iniconfig==2.3.0
openpyxl==3.1.5
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
pytest-django==4.12.0
python-decouple==3.8
qrcode==8.2
reportlab==4.4.10
sqlparse==0.5.5
